+++ spec/visual_regression_tests/all_components_spec.rb
@@ -26,7 +26,7 @@ describe "visual regression test runner Percy", :visual_regression do
     component_links = find("#list-all-components-in-the-gem")
       .all("a")
       .map do |link|
-        URI("#{link[:href]}/preview")
+        URI("#{link[:href]}/preview?percy=true")
       end
 
     # The public layout component has two previews - one with multiple versions
@@ -38,11 +38,9 @@ describe "visual regression test runner Percy", :visual_regression do
 
       visit(link)
 
-      name = title.gsub(/(: Default|) preview - Component Guide/, "")
-
       page.find(:css, "body > #wrapper", wait: 10)
 
-      page.percy_snapshot(name)
+      page.percy_snapshot(title)
     end
 
     travel_back